+++ R/utils.R
@@ -72,12 +72,13 @@ mclock <- function() .Call(rnng_clock)
 #' Sleep function. May block for longer than requested, with the actual wait
 #'     time determined by the capabilities of the underlying system.
 #'
-#' @param msec integer number of milliseconds to block the caller.
+#' @param time integer number of milliseconds to block the caller.
 #'
 #' @return Invisible NULL.
 #'
-#' @details If 'msec' is non-integer, it will be coerced to integer. Non-numeric
-#'     input will be ignored and return immediately.
+#' @details Non-integer values for 'time' are coerced to integer, and the
+#'     absolute value is taken (the sign is ignored). Non-numeric values are
+#'     ignored, causing the function to return immediately.
 #'
 #'     Note that unlike \code{\link{Sys.sleep}}, this function is not
 #'     user-interruptible by sending SIGINT e.g. with ctrl + c.
@@ -87,7 +88,7 @@ mclock <- function() .Call(rnng_clock)
 #'
 #' @export
 #'
-msleep <- function(msec) invisible(.Call(rnng_sleep, msec))
+msleep <- function(time) invisible(.Call(rnng_sleep, time))
 
 #' NNG Random Number Generator
 #'