+++ NEWS.md
@@ -1,11 +1,13 @@
-# nanonext 0.8.2.9003 (development)
+# nanonext 0.8.2.9004 (development)
 
 #### New Features
 
+* New `signal()` function signals a condition variable after a specified time (from a newly-created thread).
 * `ncurl()` gains a 'timeout' argument.
 
 #### Updates
 
+* The name of the single argument to `msleep()` has been changed to 'time' from 'msec'.
 * Functions `pipe_notify()`, `lock()` and `unlock()` now error if unsuccessful rather than returning with a warning.
 * For compiling bundled 'libmbedtls' and 'libnng' libraries from source, R's configured C compiler is now chosen over the system default where this is different.
 * Bundled 'libmbedtls' source updated to v3.4.0.